% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.single.VSM.loop.R
\name{analyze.single.VSM.loop}
\alias{analyze.single.VSM.loop}
\title{(OBSOLETE) analyzes a single hysteresis loop and return two graphs
along with essential data}
\usage{
analyze.single.VSM.loop(data)
}
\arguments{
\item{data}{hyst data frame}
}
\value{
list
}
\description{
(OBSOLETE) analyzes a single hysteresis loop and return two graphs
along with essential data
}
\examples{
filename = system.file("extdata", "20170620_BITHERMAL_SF_VSM_SF170517SI2_MVSH_3K.DAT", package="quantumPPMS")
d = ppms.load(filename)
d2 = vsm.get.HystLoops(d)
data = subset(d2, loop == 1)
m = analyze.single.VSM.loop(data)
}
