% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Mcorr.R
\name{get.Mcorr}
\alias{get.Mcorr}
\title{Removes background susceptibility}
\usage{
get.Mcorr(obj)
}
\arguments{
\item{obj}{PPMSdata object}
}
\value{
Mcorr
}
\description{
from all magnetic sweep data
}
\examples{
filename = vsm.getSampleFiles()[1]
d = ppms.load(filename)
get.Mcorr(d)
}
