% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vsm.sweepData.R
\name{get.vsm.sweepData}
\alias{get.vsm.sweepData}
\title{Return only data of magnetic sweeps (hyst loop)}
\usage{
get.vsm.sweepData(d)
}
\arguments{
\item{d}{VSM data frame}
}
\value{
VSM data frame with sweepData column
}
\description{
Return only data of magnetic sweeps (hyst loop)
}
\examples{
filename = system.file("extdata", "20170620_BITHERMAL_SF_VSM_SF170517SI2_MVSH_3K.DAT", package="quantumPPMS")
d = ppms.load(filename)
d = get.vsm.sweepData(d)
d1 = subset(d, sweepData==1)
}
