% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVSM.type.R
\name{getVSM.type}
\alias{getVSM.type}
\title{Categorize VSM measurement}
\usage{
getVSM.type(filename)
}
\arguments{
\item{filename}{name of the VSM file}
}
\value{
data frame with Tchg and Hchg
}
\description{
use this function to explore the cut-off for separating the type of
VSM file, whether a particular loop is either
MvsT or MvsH.
}
