% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSMdata.R
\name{initialize,VSMdata-method}
\alias{initialize,VSMdata-method}
\title{Constructor method of VSMdata Class}
\usage{
\S4method{initialize}{VSMdata}(
  .Object,
  time,
  T,
  H,
  M,
  Merr,
  Temp,
  dir,
  loop,
  description = "",
  sampleName = "",
  fullFilename
)
}
\arguments{
\item{.Object}{VSMdata object}

\item{time}{time in seconds}

\item{T}{temperature in Kelvin}

\item{H}{applied field in Oe}

\item{M}{magnetization in emu}

\item{Merr}{std. dev. of magnetization in emu}

\item{Temp}{temperature as factor}

\item{dir}{direction of sweeping field, +1 (more positive), -1 (more negative), 0 (not sweeping)}

\item{loop}{loop number}

\item{description}{description or note}

\item{sampleName}{sample name}

\item{fullFilename}{name of file}
}
\description{
Constructor method of VSMdata Class
}
