% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSMdata.R
\name{plot.VSMdata}
\alias{plot.VSMdata}
\title{plot VSMdata object}
\usage{
\method{plot}{VSMdata}(object, ...)
}
\arguments{
\item{object}{VSMdata object}

\item{...}{other summary parameters}
}
\value{
summary of VSMdata object
}
\description{
plot VSMdata object
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.load(filename)
plot(d)
}
\author{
Thomas Gredig
}
