% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.dat.info.R
\name{ppms.dat.info}
\alias{ppms.dat.info}
\title{OBSOLETE: Reads QD PPMS Header File Data}
\usage{
ppms.dat.info(fname)
}
\arguments{
\item{fname}{filename including path (OBSOLETE)}
}
\value{
list
}
\description{
OBSOLETE: Reads QD PPMS Header File Data
}
\examples{
filename = system.file("extdata", "20170620_BITHERMAL_SF_VSM_SF170517SI2_MVSH_3K.DAT", package="quantumPPMS")
d = ppms.dat.info(filename)
}
