% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.find.Tdep.R
\name{ppms.find.Tdep}
\alias{ppms.find.Tdep}
\title{Tries to find out the deposition temperature of the sample}
\usage{
ppms.find.Tdep(path.raw, sample.name)
}
\arguments{
\item{path.raw}{path to RAW data files}

\item{sample.name}{sample name string}
}
\value{
deposition temperature in degrees Celsius as a number
}
\description{
Tries to find out the deposition temperature of the sample
}
\examples{
ppms.find.Tdep('SF170517')
}
