% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.load.acms.R
\name{ppms.load.acms}
\alias{ppms.load.acms}
\title{Loads QD PPMS data from file}
\usage{
ppms.load.acms(filename)
}
\arguments{
\item{filename}{filename including path for PPMS file}
}
\value{
data frame with 5 data columns: time, T, H, M, Merr
}
\description{
Loads QD PPMS data from file
}
\examples{
filename = system.file("extdata", "20170620_BITHERMAL_SF_VSM_SF170517SI2_MVSH_3K.DAT", package="quantumPPMS")
d = ppms.load(filename)
print(paste('PPMS data loaded with ',nrow(d),'rows.'))
}
