% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.seqType.R
\name{ppms.seqType}
\alias{ppms.seqType}
\title{Returns characteristics of a PPMS data file.}
\usage{
ppms.seqType(Temp, Happ)
}
\arguments{
\item{Temp}{temperature vector (K)}

\item{Happ}{applied field vector (Oe)}
}
\value{
data.frame() with length, lowest/highest temperature, lowest/highest magnetic field, number of
data points for hysteresis loops and number for susceptibility, cateogory value
}
\description{
The category number can be used to check whether the
same sequence was used to generated the data.
}
\examples{
filename = system.file("extdata", "20170620_BITHERMAL_SF_VSM_SF170517SI2_MVSH_3K.DAT", package="quantumPPMS")
d = ppms.load(filename)
ppms.seqType(d$T, d$H)
}
