% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.vsm.hystLoops.R
\name{ppms.vsm.hystLoops}
\alias{ppms.vsm.hystLoops}
\title{Returns a vector that separates different hysteresis loops}
\usage{
ppms.vsm.hystLoops(H, M)
}
\arguments{
\item{hyst}{hyst data frame}
}
\value{
list
}
\description{
Returns a vector that separates different hysteresis loops
}
\examples{
filename = dir(pattern='DAT$', recursive=TRUE)[1]
d = ppms.load(filename)
d$loop = ppms.vsm.hystLoops(d$H, d$M)
}
