% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.R
\name{ppms.vsmdat.info}
\alias{ppms.vsmdat.info}
\title{Reads QD PPMS Header File Data}
\usage{
ppms.vsmdat.info(fname)
}
\arguments{
\item{fname}{filename including path}
}
\value{
list
}
\description{
Reads QD PPMS Header File Data
}
\examples{
d = ppms.vsmdat.info('ppms.dat')
}
