% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSMdata.R
\name{print.VSMdata}
\alias{print.VSMdata}
\title{print VSMdata object}
\usage{
\method{print}{VSMdata}(object, ...)

\method{print}{VSMdata}(object, ...)
}
\arguments{
\item{object}{VSMdata object}

\item{...}{other summary parameters}
}
\value{
summary of VSMdata object

summary of VSMdata object
}
\description{
print VSMdata object

plot VSMdata object
}
\examples{
filename = vsm.getSampleFiles()[1]
d = VSM.load(filename)
print(d)
filename = vsm.getSampleFiles()[1]
d = vsm.load(filename)
plot(d)
}
\author{
Thomas Gredig

Thomas Gredig
}
