% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_VSMdata.R
\name{summary.VSMdata}
\alias{summary.VSMdata}
\title{summary of VSMdata object}
\usage{
\method{summary}{VSMdata}(object, ...)
}
\arguments{
\item{object}{VSMdata object}

\item{...}{other summary parameters}
}
\value{
summary of VSMdata object
}
\description{
summary of VSMdata object
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
summary(d)
}
\author{
Thomas Gredig
}
