% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.analyzeLoop.R
\name{vsm.analyzeLoop}
\alias{vsm.analyzeLoop}
\title{Analyzes a single hysteresis loop}
\usage{
vsm.analyzeLoop(obj, loop = 1)
}
\arguments{
\item{obj}{VSMdata object}

\item{loop}{number of the hysteresis loop to display}
}
\value{
list
}
\description{
return two graphs along with essential data
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
analyze.single.VSM.loop(d)
}
