% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSMdata.R
\name{vsm.data.frame}
\alias{vsm.data.frame}
\title{Convert VSMdata object to data.frame()}
\usage{
vsm.data.frame(obj)
}
\arguments{
\item{obj}{VSMdata object}
}
\value{
data frame with VSM data
}
\description{
Convert VSMdata object to data.frame()
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
q = vsm.data.frame(d)
head(q)
ggplot(q, aes(H, M, col=loop)) + geom_point()
}
\author{
Thomas Gredig
}
