% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm_get_Coercivity.R
\name{vsm.get.Coercivity}
\alias{vsm.get.Coercivity}
\title{returns coercive field of hysteresis loop}
\usage{
vsm.get.Coercivity(H, M)
}
\arguments{
\item{H}{applied magnetic field vector (Oe)}

\item{M}{magnetization vecotr (emu)}
}
\value{
coercivity (Oe)
}
\description{
returns coercive field of hysteresis loop
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
d1 = vsm.getLoop(d, lp=1, direction=1 )
df = vsm.data.frame(d1)
vsm.get.Coercivity(df$H, df$M)
}
