% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.R
\name{vsm.get.HystLoops}
\alias{vsm.get.HystLoops}
\title{extract hysteresis loops}
\usage{
vsm.get.HystLoops(data)
}
\arguments{
\item{fname}{data frame with time, T, H, M, Merr}
}
\value{
list
}
\description{
extract hysteresis loops
}
\examples{
h = vsm.get.HystLoops(d)
}
