% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.get.Coercivity.R
\name{vsm.get.Msat}
\alias{vsm.get.Msat}
\title{returns saturation magnetization of hysteresis loops}
\usage{
vsm.get.Msat(data)
}
\arguments{
\item{data}{data frame with H, M, and T}
}
\value{
matrix with 5 columns, Msat, Msat.sd, intercept, susceptibility, Temeprature
}
\description{
returns saturation magnetization of hysteresis loops
}
\examples{
filename = vsm.getSampleFiles()
d = vsm.import(filename, dataFrame=TRUE)
vsm.get.Msat(d)
}
