% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.getLoop.R
\name{vsm.getLoop}
\alias{vsm.getLoop}
\title{Subset of VSMdata object}
\usage{
vsm.getLoop(obj, loop = 1)
}
\arguments{
\item{obj}{VSMdata object}

\item{loop}{number of the loop}
}
\value{
VSMdata object with subset of selected loop
}
\description{
Subset of VSMdata object
}
\examples{
filename = vsm.getSampleFiles()
d = ppms.load(filename)
vsm.getLoop(d)
}
