% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.getSampleFiles.R
\name{vsm.getSampleFiles}
\alias{vsm.getSampleFiles}
\title{Sample PPMS sample data files}
\usage{
vsm.getSampleFiles(type = "all")
}
\arguments{
\item{type}{can be "all" or "version" for set of different version files for testing}
}
\value{
vector with path/filename to data files
}
\description{
returns sample data files with PPMS data
}
\examples{
vsm.getSampleFiles("version")
file.list = vsm.getSampleFiles()
print(paste("Found",length(file.list),"data files."))
}
\author{
Thomas Gredig
}
