% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm_getSusceptibility.R
\name{vsm.getSusceptibility}
\alias{vsm.getSusceptibility}
\title{Returns the susceptibility background (high H)}
\usage{
vsm.getSusceptibility(obj, direction = -1, singleLoop = FALSE)
}
\arguments{
\item{obj}{VSMdata object}

\item{direction}{+1 or -1 for the direction to use for fit (-1 default)}

\item{singleLoop}{if \code{TRUE}, will evaluate only one loop regardless of temperature}
}
\value{
VSM data frame with sweepData column
}
\description{
for each temperature of the magnetic field sweep data
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
d1 = vsm.getLoop(d,direction = -1)
vsm.getSusceptibility(d1, direction=-1)
}
