% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm_guessTdep.R
\name{vsm.guessTdep}
\alias{vsm.guessTdep}
\title{Tries to find out the deposition temperature}
\usage{
vsm.guessTdep(obj)
}
\arguments{
\item{obj}{VSMdata object}
}
\value{
deposition temperature in degrees Celsius as a number
}
\description{
Retrieve deposition temperature of sample
}
\examples{
d = vsm.import(vsm.getSampleFiles()[1])
vsm.guessTdep(d)
}
