% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm_help.R
\name{vsm.help}
\alias{vsm.help}
\title{VSM Help}
\usage{
vsm.help(name = "ppms.load")
}
\arguments{
\item{name}{deprecated function name}
}
\description{
Prints instructions on how to convert from obsolete functions to
new functions
}
