% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.R
\name{vsm.hyst.stats}
\alias{vsm.hyst.stats}
\title{Statistics from Hysteresis Loop}
\usage{
vsm.hyst.stats(hyst)
}
\arguments{
\item{fname}{hyst data frame}
}
\value{
list
}
\description{
Statistics from Hysteresis Loop
}
\examples{
m = vsm.hyst.stats(h)
}
