% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.hystStats.R
\name{vsm.hystStatsLoop}
\alias{vsm.hystStatsLoop}
\title{Statistics for one Hysteresis Loop}
\usage{
vsm.hystStatsLoop(obj, loop = 1, direction = 1)
}
\arguments{
\item{obj}{VSMdata object}

\item{loop}{number of the loop}

\item{direction}{+1 or -1 for direction of applied field}
}
\value{
list or NULL if not a "MvsH" type
}
\description{
Retrieve statistics for one loop, if you need to get statistics for
all loops in the VSMdata object, then use vsm.hystStat() instead
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
t(vsm.hystStatsLoop(d))

}
