% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.import.R
\name{vsm.import}
\alias{vsm.import}
\title{Import VSM data from file}
\usage{
vsm.import(filename, dataFrame = FALSE, verbose = FALSE)
}
\arguments{
\item{filename}{filename including path for VSM file}

\item{dataFrame}{if \code{TRUE}, then will return data.frame compatible with v0.2 and before}

\item{verbose}{if \code{TRUE} output additional information}
}
\value{
VSMdata object
}
\description{
Import VSM data from file
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
summary(d)
}
