% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm_info.R
\name{vsm.info}
\alias{vsm.info}
\title{VSM/MPMS file information}
\usage{
vsm.info(filename)
}
\arguments{
\item{filename}{name of the file to read}
}
\value{
one-row data.frame with file metadata
}
\description{
Reads metadata from the file header.
}
\details{
Supports:
  Newer PPMS/VSM: INFO,value,FIELDNAME
  Legacy MPMS:    INFO,FIELDNAME,value
  Legacy MPMS:    INFO,FIELDNAME: value
}
