% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.load.R
\name{vsm.load}
\alias{vsm.load}
\title{Loads QD VSM data from file}
\usage{
vsm.load(filename, dataFrame = FALSE)
}
\arguments{
\item{filename}{filename including path for VSM file}

\item{dataFrame}{if \code{TRUE}, then will return data.frame compatible with v0.2 and before}
}
\value{
VSMdata object
}
\description{
Loads QD VSM data from file
}
\examples{
filename = vsm.getSampleFiles()[1]
d = ppms.load(filename)
summary(d)
}
