% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.readHeader.R
\name{vsm.readHeader}
\alias{vsm.readHeader}
\title{Reads the VSM file header}
\usage{
vsm.readHeader(filename, returnLinesOnly = FALSE)
}
\arguments{
\item{filename}{VSM filename usually ends with .dat}
}
\description{
Reads the VSM file header
}
