% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.validFile.R
\name{vsm.version}
\alias{vsm.version}
\title{VSM DAT file}
\usage{
vsm.version(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of the file to check (include path)}

\item{verbose}{output app name}
}
\value{
version of VSM file
}
\description{
VSM DAT file
}
\examples{
filename = vsm.getSampleFiles()[1]
vsm.version(filename)

}
