% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSMdata.R
\docType{class}
\name{VSMdata-class}
\alias{VSMdata-class}
\title{VSM data class}
\description{
A S4 class to store and manipulate VSM data
}
\section{Slots}{

\describe{
\item{\code{time}}{time in seconds}

\item{\code{T}}{temperature in Kelvin}

\item{\code{H}}{applied field in Oe}

\item{\code{M}}{magnetization in emu}

\item{\code{Merr}}{std. dev. of magnetization in emu}

\item{\code{Temp}}{temperature as factor}

\item{\code{dir}}{direction of sweeping field, +1 (more positive), -1 (more negative), 0 (not sweeping)}

\item{\code{loop}}{loop number}

\item{\code{type}}{"MvsH" or "MvsT"}

\item{\code{Mcorr}}{magnetization with substrate susceptibility removed for "MvsH" type}

\item{\code{description}}{description or note}

\item{\code{sampleName}}{sample name}

\item{\code{fullFilename}}{name of file}
}}

