% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSMdata.R
\name{VSMdata}
\alias{VSMdata}
\title{Initialize the AFMdata object}
\usage{
VSMdata(
  time,
  T,
  H,
  M,
  Merr,
  Temp,
  dir,
  description = "",
  sampleName = "",
  fullFilename = ""
)
}
\arguments{
\item{time}{time in seconds}

\item{T}{temperature in Kelvin}

\item{H}{applied field in Oe}

\item{M}{magnetization in emu}

\item{Merr}{std. dev. of magnetization in emu}

\item{Temp}{temperature as factor}

\item{dir}{direction of sweeping field, +1 (more positive), -1 (more negative), 0 (not sweeping)}

\item{description}{description or note}

\item{sampleName}{sample name}

\item{fullFilename}{name of file}
}
\description{
Initialize the AFMdata object
}
