% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Mcorr.R
\name{get.Mcorr}
\alias{get.Mcorr}
\title{removes background susceptibility}
\usage{
get.Mcorr(d)
}
\arguments{
\item{d}{VSM data frame}
}
\value{
added column Mcorr with the corrected magnetization value
}
\description{
from all magnetic sweep data
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename, dataFrame=TRUE)
head(get.Mcorr(d))
}
