% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Mcorr.R
\name{get.Mcorr}
\alias{get.Mcorr}
\title{removes background susceptibility from all
magnetic sweep data}
\usage{
get.Mcorr(d)
}
\arguments{
\item{d}{VSM data frame}
}
\value{
added column Mcorr with the corrected magnetization value
}
\description{
removes background susceptibility from all
magnetic sweep data
}
\examples{
filename = vsm.getSampleFiles()
d = ppms.load(filename)
d = get.vsm.sweepData(d)
d1 = subset(d, sweepData==1)
d1$Mcorr = get.Mcorr(d1)
}
