% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vsm.susceptibility.R
\name{get.vsm.susceptibility}
\alias{get.vsm.susceptibility}
\title{Returns the susceptibility background (high H) for
each temperature of the magnetic field sweep data}
\usage{
get.vsm.susceptibility(d, my.SweepDirection = -1)
}
\arguments{
\item{d}{VSM data frame}

\item{my.SweepDirection}{+1 or -1 for the direction to use for fit (-1 default)}
}
\value{
VSM data frame with sweepData column
}
\description{
Returns the susceptibility background (high H) for
each temperature of the magnetic field sweep data
}
\examples{
filename = vsm.getSampleFiles()
d = ppms.load(filename)
d = get.vsm.sweepData(d)
d1 = subset(d, sweepData==1)
get.vsm.susceptibility(d1)
}
