% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vsm.sweepData.R
\name{get.vsm.sweepData}
\alias{get.vsm.sweepData}
\title{OBSOLETE: Return only data of magnetic sweeps (hyst loop)}
\usage{
get.vsm.sweepData(obj)
}
\arguments{
\item{obj}{PPMSdata object}
}
\value{
VSM data frame with sweepData column
}
\description{
OBSOLETE: Return only data of magnetic sweeps (hyst loop)
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
d = get.vsm.sweepData(d)
}
