% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VSMdata.R
\name{plot.VSMdata}
\alias{plot.VSMdata}
\title{plot VSMdata object}
\usage{
\method{plot}{VSMdata}(x, ...)
}
\arguments{
\item{x}{VSMdata object}

\item{...}{other summary parameters}
}
\value{
summary of VSMdata object
}
\description{
plot VSMdata object
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
plot(d)
}
\author{
Thomas Gredig
}
