% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.dat.info.R
\name{ppms.dat.info}
\alias{ppms.dat.info}
\title{Reads QD PPMS Header File Data}
\usage{
ppms.dat.info(fname)
}
\arguments{
\item{fname}{filename including path}
}
\value{
list
}
\description{
Reads QD PPMS Header File Data
}
\examples{
filename = dir(pattern='DAT$', recursive=TRUE)[1]
d = ppms.dat.info(filename)
}
