% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.dat.info.R
\name{ppms.dat.info2}
\alias{ppms.dat.info2}
\title{Reads QD PPMS Header File Data (General)
also returns the PPMS option ("VSM","ACMS", etc.)}
\usage{
ppms.dat.info2(fname)
}
\arguments{
\item{fname}{filename including path}
}
\value{
data frame
}
\description{
Reads QD PPMS Header File Data (General)
also returns the PPMS option ("VSM","ACMS", etc.)
}
\examples{
filename = dir(pattern='DAT$', recursive=TRUE)[1]
d = ppms.dat.info2(filename)
}
