% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.load.acms.R
\name{ppms.load.acms}
\alias{ppms.load.acms}
\title{Loads QD PPMS ACMS data from file}
\usage{
ppms.load.acms(filename)
}
\arguments{
\item{filename}{filename including path for PPMS file}
}
\value{
data frame with 9 data columns: time, T, M, freq, Ampl, Mdc, M1, M2, Mtot
}
\description{
Loads QD PPMS ACMS data from file
}
\examples{
d = ppms.load.acms(filename)
print(paste('PPMS data loaded with ',nrow(d),'rows.'))
}
