% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.load.R
\name{ppms.load}
\alias{ppms.load}
\title{Loads QD PPMS data from file}
\usage{
ppms.load(filename)
}
\arguments{
\item{filename}{filename including path for PPMS file}
}
\value{
data frame with 5 data columns: time, T, H, M, Merr
}
\description{
Loads QD PPMS data from file
}
\examples{
filename = dir(pattern='DAT$', recursive=TRUE)[1]
d = ppms.load(filename)
}
