% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.R
\name{ppms.load}
\alias{ppms.load}
\title{Loads QD PPMS data from file}
\usage{
ppms.load(fname)
}
\arguments{
\item{fname}{filename including path}
}
\value{
data frame with 5 data columns: time, T, H, M, Merr
}
\description{
Loads QD PPMS data from file
}
\examples{
d = ppms.load('ppms.dat')
}
