% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.quickLoad.R
\name{ppms.quickLoad}
\alias{ppms.quickLoad}
\title{Loads several Quantum Design PPMS files at once.}
\usage{
ppms.quickLoad(ppms.file.list)
}
\arguments{
\item{ppms.file.list}{vector with full path, filenames}
}
\value{
data frame with 7 data columns: time, T, H, M, Merr, type, no
}
\description{
It will add a column with "no" numbering each file
The "type" column returns "hyst" for hysteresis loops
and "temp" for dc susceptiblility (M vs. T)
}
\examples{
file.list = c(system.file("extdata", "20170620_BITHERMAL_SF_VSM_SF170517SI2_MVSH_3K.DAT", package="quantumPPMS"))
d = ppms.quickLoad(file.list)
}
