% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.quickLoad.R
\name{ppms.quickLoad}
\alias{ppms.quickLoad}
\title{Loads several Quantum Design PPMS files at once.}
\usage{
ppms.quickLoad(ppms.file.list, verbose = TRUE)
}
\arguments{
\item{ppms.file.list}{vector with full path, filenames}

\item{verbose}{if \code{TRUE} additional information}
}
\value{
data frame with 7 data columns: time, T, H, M, Merr, type, no
}
\description{
It will add a column with "no" numbering each file
The "type" column returns "hyst" for hysteresis loops
and "temp" for dc susceptiblility (M vs. T)
}
\examples{
file.list = vsm.getSampleFiles(type = 'all')[1]
d = ppms.quickLoad(file.list, verbose = FALSE)
}
