% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppms.vsm.hystLoops.R
\name{ppms.vsm.hystLoops}
\alias{ppms.vsm.hystLoops}
\title{Returns a vector that separates different hysteresis loops}
\usage{
ppms.vsm.hystLoops(H, M)
}
\arguments{
\item{H}{applied magnetic field (Oe)}

\item{M}{magnetization (emu)}
}
\value{
list
}
\description{
Returns a vector that separates different hysteresis loops
}
\examples{
filename = system.file("extdata", "20170620_BITHERMAL_SF_VSM_SF170517SI2_MVSH_3K.DAT", package="quantumPPMS")
d = ppms.load(filename)
d$loop = ppms.vsm.hystLoops(d$H, d$M)
head(d)
}
