% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSMdata.R
\name{print.VSMdata}
\alias{print.VSMdata}
\title{print VSMdata object}
\usage{
\method{print}{VSMdata}(x, ...)
}
\arguments{
\item{x}{VSMdata object}

\item{...}{other summary parameters}
}
\value{
summary of VSMdata object
}
\description{
print VSMdata object
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
print(d)
}
\author{
Thomas Gredig
}
