% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{quantumPPMS-deprecated}
\alias{quantumPPMS-deprecated}
\alias{get.vsm.sweepData}
\alias{ppms.getSusceptibility}
\alias{ppms.removeSubstrateMagnetization}
\alias{ppms.seqType}
\alias{ppms.vsm.hystLoops}
\alias{vsm.get.HystLoops}
\alias{vsm.stats}
\title{Deprecated functions of quantumPPMS}
\usage{
get.vsm.sweepData(obj)

ppms.getSusceptibility(h)

ppms.removeSubstrateMagnetization(hyst)

ppms.seqType(Temp, Happ)

ppms.vsm.hystLoops(H, M)

vsm.get.HystLoops(data)

vsm.stats(hyst)
}
\arguments{
\item{obj}{PPMSdata object}

\item{Temp}{temperature vector (K)}

\item{Happ}{applied field vector (Oe)}

\item{H}{applied magnetic field (Oe)}

\item{M}{magnetization (emu)}

\item{data}{data frame with time, T, H, M, Merr}
}
\description{
get.vsm.sweepData,ppms.removeSubstrateMagnetization and ppms.getSusceptibility are depreacted, so avoid
using this or downgrade to an older function; these functions may disappear
from this package
}
\details{
'get.vsm.sweepData' Return only data of magnetic sweeps (hyst loop)

'ppms.seqType': Returns characteristics of a PPMS data file.

'ppms.vsm.hystLoops': Returns a vector that separates different hysteresis loops
'vsm.get.HystLoops' extract hysteresis loops (OBSOLETE, use ppms.vsm.hystLoops)

use summary(PPMSdata) instead

The category number can be used to check whether the
same sequence was used to generated the data.
}
