% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.get.Coercivity.R
\name{vsm.get.Coercivity}
\alias{vsm.get.Coercivity}
\title{returns coercive field of hysteresis loop}
\usage{
vsm.get.Coercivity(H, M)
}
\arguments{
\item{H}{applied magnetic field vector (Oe)}

\item{M}{magnetization vecotr (emu)}
}
\value{
coercivity (Oe)
}
\description{
returns coercive field of hysteresis loop
}
\examples{
filename = vsm.getSampleFiles()
d = vsm.import(filename, dataFrame = TRUE)
d$loop = vsm.get.hyst.loop(d$H, d$M)
d1 = subset(d, loop==1)
vsm.get.Coercivity(d1$H, d1$M)
}
