% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.get.hyst.loop.R
\name{vsm.get.HystLoops}
\alias{vsm.get.HystLoops}
\title{extract hysteresis loops (OBSOLETE, use ppms.vsm.hystLoops)}
\usage{
vsm.get.HystLoops(data)
}
\arguments{
\item{data}{data frame with time, T, H, M, Merr}
}
\value{
list
}
\description{
extract hysteresis loops (OBSOLETE, use ppms.vsm.hystLoops)
}
\examples{
filename = dir(pattern='DAT$', recursive=TRUE)[1]
d = ppms.load(filename)
h = vsm.get.HystLoops(d)
}
