% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.getLoop.R
\name{vsm.getLoop}
\alias{vsm.getLoop}
\title{Subset of VSMdata object}
\usage{
vsm.getLoop(obj, loop = 1, direction = 1)
}
\arguments{
\item{obj}{VSMdata object}

\item{loop}{number of the loop}

\item{direction}{+1 or -1 for direction of applied field}
}
\value{
VSMdata object with subset of selected loop
}
\description{
Subset of VSMdata object
}
\examples{
filename = vsm.getSampleFiles()
d = vsm.import(filename)
plot(d)
d1 = vsm.getLoop(d)
plot(d1)
}
