% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.getSusceptibility.R
\name{vsm.getSusceptibility}
\alias{vsm.getSusceptibility}
\title{Returns the susceptibility background (high H)}
\usage{
vsm.getSusceptibility(obj, sweepDirection = -1)
}
\arguments{
\item{obj}{VSMdata object}

\item{sweepDirection}{+1 or -1 for the direction to use for fit (-1 default)}
}
\value{
VSM data frame with sweepData column
}
\description{
for each temperature of the magnetic field sweep data
}
\examples{
filename = vsm.getSampleFiles()
d = vsm.import(filename, dataFrame=TRUE)
d1 = vsm.getLoop(d,1)
vsm.getSusceptibility(d1)
}
