% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.hyst.stats.R
\name{vsm.hyst.stats}
\alias{vsm.hyst.stats}
\title{Statistics from Hysteresis Loop}
\usage{
vsm.hyst.stats(hyst)
}
\arguments{
\item{hyst}{hyst data frame}
}
\value{
list
}
\description{
Statistics from Hysteresis Loop
}
\examples{
filename = vsm.getSampleFiles()
d = ppms.load(filename)
h = vsm.get.HystLoops(d)
m = vsm.hyst.stats(h)
}
