% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.hystStats.R, R/vsm_hystStats.R
\name{vsm.hystStats}
\alias{vsm.hystStats}
\title{Statistics for all Loops in VSM object}
\usage{
vsm.hystStats(obj)

vsm.hystStats(obj)
}
\arguments{
\item{obj}{VSMdata object}
}
\value{
list

list
}
\description{
This function loops through all hyst loops in VSMdata
using vsm.hystStats() and adds those parameters

This function loops through all hyst loops in VSMdata
using vsm.hystStats() and adds those parameters
}
\examples{
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
t(vsm.hystStats(d))
filename = vsm.getSampleFiles()[1]
d = vsm.import(filename)
t(vsm.hystStats(d))
}
\author{
Thomas Gredig
}
