% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.hystStats.R
\name{vsm.hystStats}
\alias{vsm.hystStats}
\title{Statistics from Hysteresis Loop}
\usage{
vsm.hystStats(obj, loop = 1, direction = 1)
}
\arguments{
\item{obj}{VSMdata object}

\item{loop}{number of the loop}

\item{direction}{+1 or -1 for direction of applied field}
}
\value{
list
}
\description{
Statistics from Hysteresis Loop
}
\examples{
filename = vsm.getSampleFiles()
d = vsm.load(filename)
vsm.hystStats(d)
}
