% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.info.R
\name{vsm.info}
\alias{vsm.info}
\title{Reads QD PPMS Header File Data (General)}
\usage{
vsm.info(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame
}
\description{
also returns the PPMS option ("VSM","ACMS","LogData","Resistivity")
}
\examples{
filename = vsm.getSampleFiles()
vsm.info(filename)
}
