% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm_info.R
\name{vsm.info}
\alias{vsm.info}
\title{Reads VSM data file header}
\usage{
vsm.info(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame
}
\description{
also returns the PPMS option (VSM,ACMS,LogData,Resistivity)
}
\examples{
filename = vsm.getSampleFiles()[1]
vsm.info(filename)
}
