% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.validFile.R
\name{vsm.validFile}
\alias{vsm.validFile}
\title{VSM DAT file}
\usage{
vsm.validFile(filename)
}
\value{
\code{TRUE} if it is a MPMS or VSM DAT file
}
\description{
VSM DAT file
}
\examples{
filename = vsm.getSampleFiles()[1]
vsm.validFile(filename)
}
