% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm_validFile.R
\name{vsm.validFile}
\alias{vsm.validFile}
\title{Check for valid VSM file}
\usage{
vsm.validFile(filename)
}
\arguments{
\item{filename}{name of the file to check (include path)}
}
\value{
\code{TRUE} if it is a MPMS or VSM DAT file
}
\description{
A valid VSM data file exists and needs to contain data, so >2kB in size, it also should have VSM or MPMS in
the header
}
\examples{
filename = vsm.getSampleFiles()[1]
vsm.validFile(filename)
}
