% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsm.validFile.R
\name{vsm.version}
\alias{vsm.version}
\title{VSM file version}
\usage{
vsm.version(filename, verbose = FALSE)
}
\arguments{
\item{filename}{name of the file to check (include path)}

\item{verbose}{output app name}
}
\value{
version of VSM file
}
\description{
reads the version as a number, if there are a different builds, those are added in the 3rd and 4th digit of
the number; i.e. 1.3702 for 1.37 Build 2
}
\examples{
filename = vsm.getSampleFiles()
sapply(filename, vsm.version)

}
